Moved thread local var classes into own files.
